# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SRA Resource Management System API
  version: 1.0.0
  description: >
    Business-layer API for the SRA Resource Management System (SRA-RMS).
    Provides CRUD over Clients, Projects, Resources, and Allocations, plus
    dashboard and reporting endpoints. Authentication is delegated to
    Microsoft Active Directory (OAuth2 / OpenID Connect); authorization is
    role-based (Administrator, General, Report).
  contact:
    name: SRA
    email: umesh.kodippili@sra.com.au
  license:
    name: Proprietary - SRA
servers:
  - url: https://api.rms.sra.com.au/v1
    description: Production
  - url: https://api.staging.rms.sra.com.au/v1
    description: Staging
  - url: http://localhost:5000/v1
    description: Local development
security:
  - oauth2: []
...
paths:
...
  /reports/utilisation:
    get:
      tags: [Reports]
      summary: Resource utilisation report
      operationId: getUtilisationReport
      description: Roles - Report.
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: department
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/ReportFormat'
      responses:
        '200':
          description: Utilisation report (JSON or CSV per format).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UtilisationReport' }
            text/csv:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/allocation:
    get:
      tags: [Reports]
      summary: Allocation report by project and resource
      operationId: getAllocationReport
      description: Roles - Report.
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ReportFormat'
      responses:
        '200':
          description: Allocation report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AllocationReport' }
            text/csv:
              schema: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/budget:
    get:
      tags: [Reports]
      summary: Budget consumption report
      operationId: getBudgetReport
      description: Roles - Report.
      parameters:
        - name: clientId
          in: query
          schema: { type: string, format: uuid }
        - $ref: '#/components/parameters/ReportFormat'
      responses:
        '200':
          description: Budget consumption report.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BudgetReport' }
            text/csv:
              schema: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

  # ---------------- Reference data ----------------
  /reference/{collection}:
    parameters:
      - name: collection
        in: path
        required: true
        schema:
          type: string
          enum: [departments, locations, jobTitles, skills, resourceStatuses]
    get:
      tags: [Reference data]
      summary: List reference data values
      operationId: listReferenceData
      description: Roles - Administrator, General.
      responses:
        '200':
          description: Reference data values.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ReferenceItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Reference data]
      summary: Add a reference data value
      operationId: createReferenceData
      description: Roles - Administrator.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReferenceItemCreate' }
      responses:
        '201':
          description: Reference value created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReferenceItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: Microsoft Active Directory / Entra ID (OpenID Connect).
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            api://sra-rms/access_as_user: Access the SRA-RMS API as the signed-in user.
...
  parameters:
...
    ReportFormat:
      name: format
      in: query
      description: Output format.
      schema: { type: string, enum: [json, csv], default: json }
...
  responses:
    BadRequest:
      description: The request was invalid.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: The authenticated user lacks the required role.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: The request conflicts with the current state (e.g. duplicate or dependents exist).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    UtilisationReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        rows:
          type: array
          items:
            type: object
            properties:
              resourceId: { type: string, format: uuid }
              resourceName: { type: string }
              department: { type: string }
              availableHours: { type: number, format: double }
              allocatedHours: { type: number, format: double }
              utilisation: { type: number, format: double }
    AllocationReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        rows:
          type: array
          items:
            type: object
            properties:
              projectId: { type: string, format: uuid }
              projectName: { type: string }
              clientName: { type: string }
              resourceId: { type: string, format: uuid }
              resourceName: { type: string }
              allocatedHours: { type: number, format: double }
              billable: { type: boolean }
    BudgetReport:
      type: object
      properties:
        rows:
          type: array
          items:
            type: object
            properties:
              clientName: { type: string }
              projectId: { type: string, format: uuid }
              projectName: { type: string }
              budget: { type: number, format: double }
              remaining: { type: number, format: double }
              consumed: { type: number, format: double }
              percentConsumed: { type: number, format: double }

    # ---- Reference data ----
    ReferenceItem:
      type: object
      required: [id, value]
      properties:
        id: { type: string, format: uuid }
        value: { type: string }
        active: { type: boolean, default: true }
    ReferenceItemCreate:
      type: object
      required: [value]
      properties:
        value: { type: string, minLength: 1 }